fix bug#1-1,#1-2 and decrease the results from source website to aviod non-related images.
