+++ .gitignore
@@ -2,4 +2,4 @@ ssl/
 config/
 node_modules/
 nohup.out
-
+text.js
