+++ app.js
@@ -1,111 +1,20 @@
 const express = require('express');
 const app = express();
-const superagent = require('superagent');
-const charset = require('superagent-charset');
-charset(superagent);
-const cheerio = require('cheerio');
-const request = require('request');
-const Nightmare = require('nightmare');
-const nightmare = Nightmare({ show: false, waitTimeout: 6000 });
-const nightmareHelper = require("nightmare-helper");
 const mysql = require("./util/mysqlcon.js");
-const crypto = require('crypto');
 const bodyParser = require('body-parser');
-const search = require("./util/search.js");
-const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
 const config = require('./config/development_config');
-const path = require('path');
 const secret = config.config.secret;
 const fs = require('fs');
-var privateKey  = fs.readFileSync(__dirname + '/ssl/private.key');
-var certificate = fs.readFileSync(__dirname + '/ssl/certificate.crt');
-//var chain = fs.readFileSync(__dirname + '/ssl/ca_bundle.crt');
-var credentials = { key: privateKey, cert: certificate };
-const multer = require('multer');
-const AWS = require('aws-sdk');
-const storage = multer.memoryStorage();
-const upload = multer({ storage: storage });
-const utf8 = require('utf8');
-const https = require('https');
 
+const privateKey  = fs.readFileSync(__dirname + '/ssl/private.key');
+const certificate = fs.readFileSync(__dirname + '/ssl/certificate.crt');
+const https = require('https');
+const credentials = { key: privateKey, cert: certificate };
 
 app.use(bodyParser.urlencoded({extended: false}));
 app.use(bodyParser.json());
 app.use(express.static('public'));
-
-
-
-
-app.post('/upload_user_pic',upload.single('upload_pic'), function(req, res) {
-  const time = Math.floor(Date.now() / 1000);
-  let token = req.headers.authorization.slice(7);  
-  console.log(req.file);
-  console.log(token);
-  jwt.verify(token, secret, function (err, decoded) {
-    if (err) {
-      // wrong password
-      tokenResult = false;
-      res.send({status:403,note:"Wrong token."});
-    } 
-    else if (decoded.exp <= time) {
-      //token expired
-      tokenResult = false;
-      res.send({status:408,note:"Session expired."});
-    } 
-    else{
-      let tokenResult = decoded.data;  //email
-      let query_userId = 'SELECT `user_id` FROM `user` WHERE `email` = "'+tokenResult+'";';
-      mysql.pool.getConnection(function(error, connection) {
-        if(error){
-          console.log(error);
-          res.send({error:"Error in connection database."});
-        }
-        connection.beginTransaction(function(error){
-          if(error){
-              connection.rollback();
-              connection.release(); 
-              console.log({error:"Error to begin transaction."});
-          }
-          connection.query(query_userId, function(error, result, fields){
-            if(error){
-              console.log(error);
-              res.send({error:"Query image_data Error"});
-            }
-            else{
-              connection.commit(function(error){
-                if(error){
-                  console.log({error:"Database Query Error"});
-                  return mysql.con.rollback(function(){
-                    throw error;
-                  });
-                }
-                else{
-                  let myDate = new Date();
-                  let date = myDate.getTime();
-                  let userId = result[0].user_id;
-                  let file_name = 'id_'+userId+'_profile_picture_'+date;
-                  //上傳到S3 並儲存圖片網址到資料庫
-                  uploadToS3(req.file.buffer,file_name,userId).then (function(result){
-                    res.send({status:200,note:"Update photo succesd."});
-                  });
-                  console.log("Insert data successed!");  
-                  connection.release();
-                }
-              });
-            }
-          });
-        });
-      });
-    }
-  });  
-});
-
-
-
-
-
-
 
 app.post('/edit_like_image', function(req, res) {
   const time = Math.floor(Date.now() / 1000);
@@ -121,7 +30,8 @@ app.post('/edit_like_image', function(req, res) {
       //token expired
       tokenResult = false;
       res.send({status:408,note:"Session expired."});
-    } else{
+    } 
+    else{
       //login successed
       tokenResult = decoded.data; //email
       if(list.length>0){
@@ -131,8 +41,7 @@ app.post('/edit_like_image', function(req, res) {
           let insert_image_like = 'INSERT INTO image_like(`image_id`,`email`)VALUES  ?';
           let delete_image_like = 'DELETE FROM image_like WHERE (image_id,email) IN (?)';
 
-          if(like_list.length>0 && unlike_list.length>0){ //在同一個頁面同時點了喜歡跟不喜歡
-            
+          if(like_list.length>0 && unlike_list.length>0){ //在同一個頁面同時點了喜歡跟不喜歡     
             mysql.pool.getConnection(function(error, connection) {
               if(error){
                 console.log(error);
@@ -144,27 +53,22 @@ app.post('/edit_like_image', function(req, res) {
                   console.log(error);
                   res.send({error:"Query image_like Error"});
                 }
-                else{
-                  console.log("Added like_list success:");
-                  console.log(like_list);
-                  connection.query(delete_image_like, [unlike_list] , function(error, result, fields){
+                else{                  
+                  console.log("Added like_list success:"+like_list);
+                  connection.query(delete_image_like, [unlike_list] , function(error, result, fields){                   
                     if(error){
                       console.log(error);
                       res.send({error:"Query image_like Error"});
                     }
-                    else{
-                      
-                      console.log("Delete unlike_list success:");
-                      console.log(unlike_list);
+                    else{                    
+                      console.log("Delete unlike_list success:"+unlike_list);
                       res.send({status:200, note:"Modify like_list success"});
                     }
                   }); 
                 }
               });
             });
-
           }
-
           else if(like_list.length>0 && unlike_list.length==0){
             mysql.pool.getConnection(function(error, connection) {
               if(error){
@@ -178,9 +82,7 @@ app.post('/edit_like_image', function(req, res) {
                   res.send({error:"Query image_like Error"});
                 }
                 else{
-                  console.log("Added like_list success");
-                  console.log(like_list);
-                  
+                  console.log("Added like_list success:"+like_list);
                 }
               });
             });
@@ -198,8 +100,7 @@ app.post('/edit_like_image', function(req, res) {
                   res.send({error:"Query image_like Error"});
                 }
                 else{
-                  console.log("Delete unlike_list success:");
-                  console.log(unlike_list);
+                  console.log("Delete unlike_list success:"+unlike_list);
                   res.send({status:200, note:"Delete like_list success"});
                 }
               }); 
@@ -236,290 +137,36 @@ app.get('/hot_image', function(req, res) {
 });
 
 
-
-app.post('/signup', function(req, res) {
-  let email = req.body.email;
-  let name = req.body.name;
-  let password = bcrypt.hashSync(req.body.password, 10);
-  let insert={
-    email:req.body.email,
-    password : password,
-    provider: 'native',
-    name:name
-  };
-
-  let insert_user_data = "insert into user SET ?";
-  let query_user_data = "SELECT `email` FROM user WHERE `email`='"+email+"';"
-    mysql.pool.getConnection(function(error, connection) {
-      if(error){
-        console.log(error);
-        res.send({error:"Error in connection database."});
-      }
-      connection.query(query_user_data, function(error, repeat_results, fields){
-        if(error){
-          console.log(error);
-          console.log({error:"Add user_data Error"});
-          connection.rollback(function() {
-            throw error;
-          });
-        }else if (repeat_results.length > 0) {
-          res.send('信箱已註冊！');
-        }
-        else{
-          connection.query(insert_user_data, insert, function(error, results, fields){
-            connection.release();
-            if(error){
-              console.log(error);
-              console.log({error:"Add user_data Error"});
-              connection.rollback(function() {
-                throw error;
-              });
-            }
-            else{
-              let token = jwt.sign({
-                algorithm: 'HS256',
-                exp: Math.floor(Date.now() / 1000) + (60 * 60), // token一個小時後過期。
-                data: email
-                }, secret);
-                res.setHeader('Authorization', 'Bearer '+token);
-                res.cookie('Authorization', token);
-                res.redirect("/");
-                // res.json({
-                //     result: {
-                //         status: "註冊成功。",
-                //         loginMember: "歡迎 " + name + " 的登入！"
-                //     }
-                // });
-            }
-          });
-        }
-      });
-    });
-});
-
-
-
-
-
-
-app.post('/signin', function(req, res) {
-  console.log(req.body);
-  let email = req.body.email;
-  let password = req.body.password;
-
-  let query_user_data = "SELECT * FROM user WHERE `email`='"+email+"';"
-  mysql.pool.getConnection(function(error, connection) {
-    if(error){
-      res.send({error:"Error in connection database."});
-    }
-    connection.query(query_user_data, function(error, result, fields){
-      connection.release();
-      if(error){
-        console.log({error:"Add user_data Error"});
-        connection.rollback(function() {
-          throw error;
-        });
-      }
-      else if (result.length > 0) { 
-          let compare_result = bcrypt.compareSync(password, result[0].password);
-          if(compare_result){
-            let token = jwt.sign({
-              algorithm: 'HS256',
-              exp: Math.floor(Date.now() / 1000) + (60 * 60), // token一個小時後過期。
-              data: email
-              }, secret);
-            res.setHeader('Authorization', 'Bearer '+token);
-            res.cookie('Authorization', token);
-            res.redirect('/');
-          }
-          else{
-            res.send('密碼錯誤！');
-          }
-      }
-      else{
-        res.send('此信箱尚未註冊！');
-      }
-    });
-  });
-
-});
-
-app.post('/fb_signin', function(req, res) {
-  let email = req.body.email;
-  let name = req.body.name;
-  let fb_token = req.body.access_token;
-  let query_user_data = "SELECT * FROM user WHERE `email`='"+email+"';"
-  let insert_user_data = "insert into user SET ?";
-  // let insert = {
-  //   email:email,
-  //   provider:'Facebook',
-  //   name:name,
-  // };
-  get_FB_photo(fb_token).then(function(result){
-    let insert={
-      email:email,
-      provider:'Facebook',
-      name:name,
-      user_photo : result
-    };
-    mysql.pool.getConnection(function(error, connection) {
-      if(error){
-        console.log(error);
-        res.send({error:"Error in connection database."});
-      }
-      connection.query(query_user_data, function(error, result, fields){
-        if(error){
-        console.log(error);
-        console.log({error:"Add user_data Error"});
-        connection.rollback(function() {
-          throw error;
-        });
-        }if (result.length == 0) { 
-        console.log("尚未註冊");
-        console.log(insert);
-        connection.query(insert_user_data, insert, function(error, result, fields){
-          connection.release();
-          if(error){
-            console.log(error);
-            console.log({error:"Add user_data Error"});
-            connection.rollback(function() {
-              throw error;
-            });
-          }
-          else{
-            
-            let token = jwt.sign({
-              algorithm: 'HS256',
-              exp: Math.floor(Date.now() / 1000) + (60 * 60), // token一個小時後過期。
-              data: email
-              }, secret);
-            res.setHeader('Authorization', 'Bearer '+token);
-            res.cookie('Authorization', token);
-            res.send('Sign up successed!');
-          }
-        });
-        }
-        else{
-        
-        let token = jwt.sign({
-          algorithm: 'HS256',
-          exp: Math.floor(Date.now() / 1000) + (60 * 60), // token一個小時後過期。
-          data: email
-          }, secret);
-        res.setHeader('Authorization', 'Bearer '+token);
-        res.cookie('Authorization', token);
-        res.send('Login with Facebook successed!');
-        console.log("註冊過了");
-        }
-      });
-    });
-  })
-
- });
-
-
-app.post('/member', function(req, res) {
-
-  const time = Math.floor(Date.now() / 1000);
-  
-  //判斷token是否正確
-    let token = req.headers.authorization.slice(7);  
-    jwt.verify(token, secret, function (err, decoded) {
-      if (err) {
-        // wrong password
-        tokenResult = false;
-        res.send({status:403,note:"Wrong token."});
-      } 
-      else if (decoded.exp <= time) {
-        //token expired
-        tokenResult = false;
-        res.send({status:408,note:"Session expired."});
-      } else{
-        tokenResult = decoded.data;
-        let like_image_data =[];
-        
-        let query = 'SELECT * from `image_data` AS A RIGHT JOIN `image_like` AS B ON A.`image_id` = B.`image_id` INNER JOIN `user` AS D on D.`email`= B.`email` WHERE D.`email` = "'+tokenResult+'";'
-        //'SELECT * from image_data AS A INNER JOIN image_author AS C ON A.`image_id` = C.`image_id` RIGHT JOIN `image_like` AS B ON A.`image_id` = B.`image_id` where B.`email`="'+tokenResult+'";'
-        
-        mysql.pool.getConnection(function(error, connection) {
-          if(error){
-            console.log(error);
-            res.send({error:"Error in connection database."});
-          }    
-          connection.query(query, function(error, result, fields){
-            connection.release();
-            if(error){
-            console.log({error:"Add user_data Error"});
-            connection.rollback(function() {
-              throw error;
-            });
-            }
-            else if(result.length == 0){ //沒有按下喜歡的圖片就丟 user 所有資訊
-              //console.log(tokenResult);
-              let user_query = 'SELECT * from `user` WHERE `email`="'+tokenResult+'"'
-              connection.query(user_query, function(error, result, fields){
-                
-                if(error){
-                  console.log({error:"Add user_data Error"});
-                  connection.rollback(function() {
-                    throw error;
-                  });
-                }
-                else{
-                  let user_name = result[0].name;
-                  let email = result[0].email;
-                  let user_profile_pic = result[0].user_photo;
-                  res.send({status:200, user_name:user_name,user_email:email,user_pic:user_profile_pic||"null",like_image_info:"NULL"});
-                  }
-              });
-            }
-            else{
-              for(let i=0;i<result.length;i++){
-                like_image_data.push({
-                  image_id:result[i].image_id,
-                  image_url:result[i].image_url,
-                  image_source_url:result[i].image_source_url,
-                  image_like : result[i].likes,
-                  auther_name:result[i].auther_name||'null',
-                  auther_website:result[i].auther_website||'null',
-                });
-              }
-              let user_name = result[0].name;
-              let email = result[0].email;
-              let user_profile_pic = result[0].user_photo;
-              res.send({status:200, user_name:user_name,user_email:email,user_pic:user_profile_pic||"null",total_likes:result.length, like_image_info:like_image_data});
-            }
-          });
-        });
-      }
-    });
-});
-
-
 app.post('/delete_broken_image', function(req, res) {
   let broken_image_id = JSON.parse(req.body.broken_id);
   broken_list(broken_image_id).then(function(result){
-    //console.log(result[0]);
     let broken_id_list = result;
     let delete_image_data = 'DELETE FROM image_data WHERE (`image_id`) IN (?)';
+    let delete_image_like = 'DELETE FROM image_like WHERE (`image_id`) IN (?)';
     mysql.pool.getConnection(function(error, connection) {
       if(error){
         console.log(error);
-        res.send({error:"Error in connection database."});
+        console.log({error:"Error in connection database."});
       }
-      connection.query(delete_image_data, [broken_id_list] , function(error, query_result, fields){
-        connection.release();
+      connection.query(delete_image_data, [broken_id_list] , function(error, delete_image_data_result, fields){
         if(error){
           console.log(error);
-          res.send({error:"Delete image_data Error"});
+          console.log({error:"Delete image_data Error"});
         }
         else{
-          res.send("123");
-        }
-      });
+          connection.query(delete_image_like, [broken_id_list] , function(error, delete_image_like_result, fields){
+            connection.release();
+            if(error){
+              console.log(error);
+              console.log({error:"Delete image_like Error"});
+            }
+        });
+      }
     });
   });
   
+  
+});
 });
 
 
@@ -542,16 +189,7 @@ function broken_list(data){
 
 
 
-function get_FB_photo(token){
-  return new Promise((mainResolve, mainReject) => {
-    request('https://graph.facebook.com/me/picture?type=large&redirect=false&access_token='+token, 
-    (error, response, body) => {
-      let photo_info = JSON.parse(body);
-      let FB_data = photo_info.data.url;
-      return mainResolve(FB_data);
-    });
-  });
-}
+
 
 function like_unlike_list(data,email){
   return new Promise((mainResolve, mainReject) => {
@@ -568,82 +206,6 @@ function like_unlike_list(data,email){
     return mainResolve({like_list:like_list,unlike_list:unlike_list||null,email:email})
   });
 }
-
-
-
-function uploadToS3(data,filename,userId) {
-  return new Promise((mainResolve, mainReject) => {
-    AWS.config.loadFromPath('./config/S3config.json');
-    let s3 = new AWS.S3({
-      params: {
-          Bucket: "jyhsum",
-          Key: filename, //檔案名稱
-          ACL: 'public-read' //檔案權限
-      }
-    });
-    s3.upload({
-        Body: data
-    }).on('httpUploadProgress', function(evt) {   
-        console.log(evt);//上傳進度
-    }).
-    send(function(err, data) {//上傳完畢或是碰到錯誤
-      if(err){
-        console.log(err);
-      }
-      else{
-        console.log("Upload to S3 successed!");
-        let pic_url = data.Location;
-        let update_user_pic = 'UPDATE user SET `user_photo` ="'+pic_url+'" WHERE `user_id`= "'+userId+'";';
-        let update_data = [pic_url,userId];
-        mysql.pool.getConnection(function(error, connection) {
-          if(error){
-            connection.release();
-            console.log(error);
-            res.send({error:"Error in connection database."});
-          }
-          connection.beginTransaction(function(error){
-            if(error){
-                connection.rollback();
-                connection.release(); 
-                console.log({error:"Error to begin transaction."});
-            }
-            connection.query(update_user_pic, function(error, result, fields){ 
-              if(error){
-                console.log(error);
-                return mainReject ({error:"Query image_data Error"});
-              }
-              else{
-                connection.commit(function(error){
-                  if(error){
-                    console.log({error:"Database Query Error"});
-                    return mysql.con.rollback(function(){
-                      throw error;
-                    });
-                  }
-                  else{
-                      connection.release();
-                      console.log(update_data);
-                      console.log("insert data successed");
-                      console.log('changed ' + result.changedRows + ' rows');
-                      return mainResolve ({status:"Upload to S3 successed"});
-                  }
-                });
-
-              }
-            });
-          });
-        });
-      }          
-    });
-  });
-}
-
-
-
-
-
-
-
 
 
 
@@ -664,5 +226,6 @@ httpsServer.listen(443, function () {
 
 const search_route = require("./routes/search.js");
 app.use("/photo",search_route);
-
+const member_route = require("./routes/member.js");
+app.use("/member",member_route);
