+++ public/js/photo.js
@@ -191,35 +191,39 @@ function read_new_pic(nextpage) {
 
 }
 
-let nextpage = 0;
+let nextpage = 1;
 window.addEventListener('scroll', function(e) {
   let last_known_scroll_position = window.scrollY;
   let pic_column_height = document.getElementsByClassName("pic_column")[0].clientHeight;
   let loader = document.getElementsByClassName("lds-roller")[0];
   if (last_known_scroll_position+pic_column_height>document.body.clientHeight) {
     nextpage+=0.5;
-    read_new_pic(nextpage)
-    .then(function(result) {
-      let image_data = JSON.parse(result);
-      let total_page = image_data.search_result.total_page;
-      if(total_page){
-        get_pic_column_data(image_data).then(function(pic_column_data) {
-          addImgDiv("pic_column",pic_column_data.image_for_column_1,0);
-          addImgDiv("pic_column",pic_column_data.image_for_column_2,1);
-          addImgDiv("pic_column",pic_column_data.image_for_column_3,2);
-        });       
-      }
-      else{
-        loader.style.display = "none"; 
-      }
-    })
-    .then(function(){
-      check_login_status().then(function(login_status_result) {
-        if(login_status_result) {
-          show_like_status(login_status_result);
+    if(nextpage%1==0){
+      console.log(nextpage);
+      read_new_pic(nextpage)
+      .then(function(result) {
+        let image_data = JSON.parse(result);
+        let total_page = image_data.search_result.total_page;
+        if(total_page){
+          get_pic_column_data(image_data).then(function(pic_column_data) {
+            addImgDiv("pic_column",pic_column_data.image_for_column_1,0);
+            addImgDiv("pic_column",pic_column_data.image_for_column_2,1);
+            addImgDiv("pic_column",pic_column_data.image_for_column_3,2);
+          });       
         }
+        else{
+          loader.style.display = "none"; 
+        }
+      })
+      .then(function(){
+        check_login_status().then(function(login_status_result) {
+          if(login_status_result) {
+            show_like_status(login_status_result);
+          }
+        });
       });
-    });
+    }  
+
   }
 
 });
